+++ src/main/java/nextstep/ladder/model/v1/DrawStrategy.java
@@ -1,4 +1,4 @@
-package nextstep.ladder.model;
+package nextstep.ladder.model.v1;
 
 public interface DrawStrategy {
     boolean draw();
