+++ src/main/java/nextstep/ladder/model/v1/LineLadderBuilder.java
@@ -1,4 +1,4 @@
-package nextstep.ladder.model;
+package nextstep.ladder.model.v1;
 
 import nextstep.ladder.model.api.Ladder;
 import nextstep.ladder.model.api.LadderBuilder;
