+++ src/main/java/nextstep/ladder/model/v1/RandomDrawStrategy.java
@@ -1,4 +1,4 @@
-package nextstep.ladder.model;
+package nextstep.ladder.model.v1;
 
 import java.util.Random;
