+++ src/test/java/nextstep/ladder/model/AllDrawStrategy.java
@@ -1,5 +1,7 @@
 package nextstep.ladder.model;
 
+import nextstep.ladder.model.v1.DrawStrategy;
+
 public class AllDrawStrategy implements DrawStrategy {
     @Override
     public boolean draw() {
