+++ src/test/java/nextstep/ladder/model/LineLadderTest.java
@@ -1,6 +1,8 @@
 package nextstep.ladder.model;
 
 import nextstep.ladder.model.api.Ladder;
+import nextstep.ladder.model.v1.DrawStrategy;
+import nextstep.ladder.model.v1.LineLadderBuilder;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
