+++ src/test/java/nextstep/ladder/model/LinesTest.java
@@ -1,5 +1,7 @@
 package nextstep.ladder.model;
 
+import nextstep.ladder.model.v1.Line;
+import nextstep.ladder.model.v1.Lines;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
