+++ src/test/java/nextstep/ladder/model/OneByTwoDrawStrategy.java
@@ -1,5 +1,7 @@
 package nextstep.ladder.model;
 
+import nextstep.ladder.model.v1.DrawStrategy;
+
 public class OneByTwoDrawStrategy implements DrawStrategy {
     private boolean draw;
